Add delete for consultant data

+++ app/Http/Controllers/ConsultantVirtualAccountController.php
@@ -13,8 +13,14 @@ class ConsultantVirtualAccountController extends Controller
         $this->middleware('auth:consultants-api');
     }
     
-    public function destroy($id) {
+    public function destroy(Request $request, $id) {
         $data = ConsultantVirtualAccount::findOrFail($id);
+        if($request->user()->cannot('delete', $data)) {
+            return response()->json([
+                'code' => 403,
+                'message' => 'Forbidden'
+            ],403);
+        }
         $data->delete();
         return response()->json([
             'code' => 200,
